backend/iit_bombay_ai/requirements.txt
fastapi
uvicorn
pandas
numpy
scipy
scikit-learn
xgboost
groq
python-dotenv
pydantic
